alma-api/app/requirements.txt
aiocache==0.12.3
aiostream==0.7.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
async-lru==2.3.0
azure-ai-contentsafety==1.0.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.25.3
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
diskcache==5.6.3
distro==1.9.0
emoji==2.15.0
executing==2.2.1
Faker==40.11.1
fastapi==0.135.2
googleapis-common-protos==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json_repair==0.58.7
jsonpath==0.82.2
logfire==4.30.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
MeUtils==2026.3.19.17.0.5
msal==1.35.1
msal-extensions==1.3.1
nacos-sdk-python==1.0.0
nest-asyncio==1.6.0
numpy==2.4.3
openai==2.30.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==26.0
pandas==3.0.1
pika==1.3.2
pillow==12.1.1
Pillow-PIL==0.1.dev0
protobuf==6.33.6
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.12.1
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.33.0
rich==14.3.3
schedule==1.2.2
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tinytag==2.2.1
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
wget==3.2
win32_setctime==1.2.0
wrapt==1.17.3
zipp==3.23.0
